+++ NEWS.md
@@ -1,3 +1,6 @@
+* `DocumentTermMatrix` and `TermDocumentMatrix` class are now supported in `viewxl()` function.
+
+
 # ezpickr 1.1.0
 
 * "*.Rmd" file format is now supported in `pick()` using `rmarkdown::render()` function.